Vérification des threads

+++ src/vector_threads.c
@@ -8,7 +8,7 @@
 void* add_v_v_thread(void *arg) {
     thread_data_v_v *data = (thread_data_v_v *)arg;  // On "déballe" le arg
 
-    for (int i = data->start_idx; i < data->end_idx; i++) {
+    for (uint64_t i = data->start_idx; i < data->end_idx; i++) {
         data->z->values[i] = data->x->values[i] + data->y->values[i];
     }
 
@@ -18,10 +18,11 @@ void* add_v_v_thread(void *arg) {
 void* sub_v_v_thread (void *arg) {
     thread_data_v_v *data = (thread_data_v_v *)arg;
 
-    for(int i = data->start_idx; i < data->end_idx; i++) {
+    for(uint64_t i = data->start_idx; i < data->end_idx; i++) {
         data->z->values[i] = data->x->values[i] - data->y->values[i];
     }
 
     return NULL; 
 
 }
+
